+++ js/src/index.ts
@@ -1,4 +1,4 @@
-export { LogEvent, TxExecution } from '../proto/exec_pb';
+export { Event, LogEvent, TxExecution } from '../proto/exec_pb';
 export { CallTx, TxInput } from '../proto/payload_pb';
 export { BlockRange } from '../proto/rpcevents_pb';
 export { Client } from './client';
@@ -7,5 +7,15 @@ export { Address } from './contracts/abi';
 export { makeCallTx } from './contracts/call';
 export { Contract } from './contracts/contract';
 export { Result } from './convert';
-export { EndOfStream, EventStream } from './events';
+export {
+  cancelStream,
+  EventStream,
+  isEndOfStream,
+  iterateEvents,
+  listenerFor,
+  readEvents,
+  reduceEvents,
+  Signal,
+} from './events';
 export { build } from './solts/build';
+export { Caller, defaultCall, Provider } from './solts/interface.gd';