+++ js/src/solts/lib/deployer.ts
@@ -1,13 +1,16 @@
-import ts, { factory, FunctionDeclaration } from 'typescript';
+import ts, { factory, FunctionDeclaration, SyntaxKind } from 'typescript';
 import { ABI } from './abi';
+import { contractFunctionName, contractTypeName } from './contract';
 import { linkerName } from './linker';
 import { callEncodeDeploy, Provider } from './provider';
 import { getRealType, sha3, tokenizeString } from './solidity';
 import {
+  AsyncToken,
   BufferType,
   createAssignmentStatement,
   createCall,
   createParameter,
+  createPromiseOf,
   declareConstant,
   declareLet,
   ExportToken,
@@ -16,13 +19,14 @@ import {
   StringType,
 } from './syntax';
 
+export const deployName = factory.createIdentifier('deploy');
+export const deployContractName = factory.createIdentifier('deployContract');
+
 // Variable names
 const payloadName = factory.createIdentifier('payload');
 const linkedBytecodeName = factory.createIdentifier('linkedBytecode');
 const dataName = factory.createIdentifier('data');
 const clientName = factory.createIdentifier('client');
-
-export const DeployName = factory.createIdentifier('deploy');
 
 export function generateDeployFunction(
   abi: ABI.Func | undefined,
@@ -31,7 +35,6 @@ export function generateDeployFunction(
   provider: Provider,
   abiName: ts.Identifier,
 ): FunctionDeclaration {
-  const parameters = abi ? abi.inputs?.map((input) => createParameter(input.name, getRealType(input.type))) ?? [] : [];
   const output = factory.createExpressionWithTypeArguments(PromiseType, [StringType]);
 
   const statements: ts.Statement[] = [];
@@ -73,14 +76,52 @@ export function generateDeployFunction(
     undefined,
     [ExportToken],
     undefined,
-    DeployName,
+    deployName,
     undefined,
-    [
-      createParameter(clientName, provider.type()),
-      ...links.map((link) => createParameter(factory.createIdentifier(tokenizeString(link)), StringType)),
-      ...parameters,
-    ],
+    deployParameters(abi, bytecodeName, links, provider),
     output,
     factory.createBlock(statements, true),
   );
+}
+
+export function generateDeployContractFunction(
+  abi: ABI.Func | undefined,
+  bytecodeName: ts.Identifier,
+  links: string[],
+  provider: Provider,
+) {
+  const parameters = deployParameters(abi, bytecodeName, links, provider);
+  const addressName = factory.createIdentifier('address');
+  const callDeploy = factory.createAwaitExpression(
+    createCall(deployName, [
+      ...parameters.map((p) => p.name).filter((n): n is ts.Identifier => n.kind === SyntaxKind.Identifier),
+    ]),
+  );
+  return factory.createFunctionDeclaration(
+    undefined,
+    [ExportToken, AsyncToken],
+    undefined,
+    deployContractName,
+    undefined,
+    parameters,
+    createPromiseOf(factory.createTypeReferenceNode(contractTypeName)),
+    factory.createBlock([
+      declareConstant(addressName, callDeploy),
+      factory.createReturnStatement(createCall(contractFunctionName, [clientName, addressName])),
+    ]),
+  );
+}
+
+function deployParameters(
+  abi: ABI.Func | undefined,
+  bytecodeName: ts.Identifier,
+  links: string[],
+  provider: Provider,
+): ts.ParameterDeclaration[] {
+  const parameters = abi ? abi.inputs?.map((input) => createParameter(input.name, getRealType(input.type))) ?? [] : [];
+  return [
+    createParameter(clientName, provider.type()),
+    ...links.map((link) => createParameter(factory.createIdentifier(tokenizeString(link)), StringType)),
+    ...parameters,
+  ];
 }